refactor layout and content components
