+++ src/management-system-v2/components/content.module.scss
@@ -1,19 +1,20 @@
 .Main {
-  // min-height: 40vh !important;
-  height: 100%;
+  min-height: 100vh !important;
 
   .Header {
     display: flex;
     justify-content: space-between;
     padding: 0;
     align-items: center;
+    border-bottom: 1px solid #eee;
+    margin-left: -1px;
+    padding-left: 21px;
 
     .Title {
       letter-spacing: 0.2px;
+      font-size: 18px;
+      color: #5f5f5f;
       font-weight: 300;
-      margin-bottom: 0;
-      display: inline-block;
-      margin-left: 20px;
     }
 
     .Right {
@@ -26,9 +27,9 @@
 
   .Content {
     overflow: initial;
-    padding: 10px 20px;
+    padding: 40px 20px;
     max-width: 100%;
-    // background-color: white;
+    background-color: white;
 
     @media screen and (max-width: 550px) {
       padding-left: 0px;
@@ -38,11 +39,5 @@
     &.compact {
       padding: 0;
     }
-  }
-
-  .Footer {
-    text-align: center;
-    color: #9a9a9a;
-    font-weight: 300;
   }
 }
