+++ src/management-system-v2/components/content.tsx
@@ -1,17 +1,21 @@
 'use client';
 
 import styles from './content.module.scss';
-import { FC, PropsWithChildren } from 'react';
-import { Layout, theme, Typography } from 'antd';
+import { FC, PropsWithChildren, ReactNode } from 'react';
+import { Layout as AntLayout, theme, Typography, Grid, Button } from 'antd';
+import { MenuOutlined } from '@ant-design/icons';
 import cn from 'classnames';
+import Link from 'next/link';
+import Image from 'next/image';
+import HeaderActions from './header-actions';
 
 const { Title } = Typography;
 
 type ContentProps = PropsWithChildren<{
-  /** Top left title in the header. */
-  title?: string;
+  /** Top left title in the header (or custom node). */
+  title?: ReactNode;
   /** Top right node in the header. */
-  rightNode?: React.ReactNode;
+  rightNode?: ReactNode;
   /** If true, the content won't have any padding. */
   compact?: boolean;
   /** If true, the content won't have a header. */
@@ -34,39 +38,41 @@ const Content: FC<ContentProps> = ({
   noHeader = false,
   wrapperClass,
   headerClass,
-  footerClass,
   fixedHeader = false,
 }) => {
   const {
     token: { colorBgContainer },
   } = theme.useToken();
 
+  const breakpoint = Grid.useBreakpoint();
+
   return (
-    <Layout className={cn(styles.Main, wrapperClass)}>
-      <div
-        style={{
-          background: colorBgContainer,
-          // paddingTop: '7px',
-          //borderBottom: '1px solid #eee',
-          // position: fixedHeader ? 'fixed' : 'static',
-          // width: fixedHeader ? '100%' : 'auto',
-        }}
-        className={cn(styles.Header, headerClass)}
-      >
-        <Title level={3} className={styles.Title}>
-          {title}
-        </Title>
-        <div className={styles.Right}>{rightNode}</div>
-      </div>
-      <Layout.Content
-        className={cn(styles.Content, { [styles.compact]: compact })}
-        style={{
-          background: colorBgContainer,
-        }}
-      >
+    <AntLayout className={cn(styles.Main, wrapperClass)}>
+      {noHeader ? null : (
+        <AntLayout.Header className={styles.Header}>
+          <div className={styles.Title}>{title}</div>
+          {breakpoint.xs ? (
+            // Hamburger menu for mobile view
+            <>
+              <Button
+                type="text"
+                size="large"
+                style={{ marginTop: '20px', marginLeft: '15px' }}
+                icon={<MenuOutlined style={{ fontSize: '170%' }} />}
+              />
+            </>
+          ) : (
+            // Logout and User Profile in header for screens larger than 412px
+            <>
+              <HeaderActions />
+            </>
+          )}
+        </AntLayout.Header>
+      )}
+      <AntLayout.Content className={cn(styles.Content, { [styles.compact]: compact })}>
         {children}
-      </Layout.Content>
-    </Layout>
+      </AntLayout.Content>
+    </AntLayout>
   );
 };
