+++ src/management-system-v2/package.json
@@ -20,27 +20,27 @@
     "singleQuote": true
   },
   "dependencies": {
-    "@monaco-editor/react": "^4.5.2",
-    "@tanstack/react-query": "4.35.3",
-    "antd": "5.9.3",
+    "@monaco-editor/react": "4.5.2",
+    "@tanstack/react-query": "4.35.7",
+    "antd": "5.9.4",
     "bpmn-js": "13.2.0",
     "bpmn-js-differ": "2.0.2",
     "classnames": "2.3.2",
-    "monaco-editor": "^0.41.0",
-    "next": "13.5.3",
+    "monaco-editor": "0.43.0",
+    "next": "13.5.4",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "zustand": "4.4.1",
-    "immer": "10.0.2",
+    "zustand": "4.4.2",
+    "immer": "10.0.3",
     "fuse.js": "6.6.2",
     "@casl/ability": "6.5.0",
-    "openapi-fetch": "0.6.1",
+    "openapi-fetch": "0.7.8",
     "@proceed/bpmn-helper": "1.0.0"
   },
   "devDependencies": {
-    "@tanstack/eslint-plugin-query": "4.34.1",
-    "@tanstack/react-query-devtools": "4.35.3",
-    "@types/node": "20.8.0",
+    "@tanstack/eslint-plugin-query": "4.36.0",
+    "@tanstack/react-query-devtools": "4.35.7",
+    "@types/node": "20.8.2",
     "@types/react": "18.2.24",
     "@types/react-dom": "18.2.8",
     "cross-env": "7.0.3",
@@ -50,7 +50,7 @@
     "prettier": "3.0.3",
     "ts-node": "10.9.1",
     "typescript": "5.2.2",
-    "openapi-typescript": "6.3.2",
+    "openapi-typescript": "6.7.0",
     "sass": "1.68.0"
   }
 }
